Add: Adds first solution
This is a first solution draft for the linkedin task at Coding Bootcamps Europe. It kind of works and almost everything works, but the code at this state is highly inefficient. Whenever a card is being deleted, all cards are being rendered again. Also, by doing this, the states of the 'Pending' buttons change to their default value ('Connect'), while the pending counter remains the same.