Fix: Fixes inefficient code
The code now runs much more efficient by only rendering those elements that are not rendered yet. By clicking on the 'delete' button, the chosen element is being removed from the array and DOM and only this element is being replaced and rendered again. This opened the door for a new bug however: The 'Connect' and 'Pending' buttons work just fine, until an element is being removed. Then the connect button does not change its text anymore and the pending count goes down, even below zero.